--- a/test/GRAPHFILES
+++ b/test/GRAPHFILES
@@ -327,3 +327,5 @@ statements/lydia/moduleVersusFunction.graph
 # suite: Samples (bogus)
 Samples/Performance/foo.graph
 Samples/Performance/cleanCopy/foobar.graph
+# suite: Lifetime Management
+memory/lifetime/allocation.graph