--- a/test/memory/lifetime/allocation.execopts
+++ b/test/memory/lifetime/allocation.execopts
@@ -1,0 +1,1 @@
+--printTiming=false --allocationsPerTrial=10