--- a/test/memory/lifetime/allocation.graph
+++ b/test/memory/lifetime/allocation.graph
@@ -1,5 +1,5 @@
 perfkeys: Unmanaged-Time:, Shared-Time:, Owned-Time:
 files: allocation.dat, allocation.dat, allocation.dat
 graphkeys: Unmanaged, Shared, Owned
-ylabel: Time (milliseconds)
-graphtitle: Lifetime Management Overhead (Unmanaged as Baseline)
+ylabel: Time (seconds)
+graphtitle: Serial 'object' Allocation