--- a/test/memory/lifetime/allocation.perfexecopts
+++ b/test/memory/lifetime/allocation.perfexecopts
@@ -1,0 +1,1 @@
+--printTiming