--- a/test/memory/lifetime/allocation.perfexecopts
+++ b/test/memory/lifetime/allocation.perfexecopts
@@ -1,1 +1,0 @@
---printTiming