--- a/test/memory/lifetime/allocation.perfkeys
+++ b/test/memory/lifetime/allocation.perfkeys
@@ -1,0 +1,3 @@
+Unmanaged-Time:
+Shared-Time:
+Owned-Time: