--- a/test/memory/lifetime/allocation.prediff
+++ b/test/memory/lifetime/allocation.prediff
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+
+print(0);