--- a/test/memory/lifetime/allocation.prediff
+++ b/test/memory/lifetime/allocation.prediff
@@ -1,3 +1,0 @@
-#!/usr/bin/env python
-
-print(0);