This Pull Request adds a test for serial shared-memory allocations of
`shared`/`owned`/`unmanaged` to profile the performance overheads of each.

This shows that `shared` is ~2.5x slower and that there's overhead for `owned`
under `--no-local`:

Timings on Louis's Mac-book:

| config     | Unmanaged | Shared  | Owned   |
| ---------- | --------- | ------- | ------- |
| local      | 0.919s    | 2.480s  | 0.900s  |
| --no-local | 0.938s    | 2.782s  | 1.362s  |

Timings on chapcs07:

| config     | Unmanaged | Shared  | Owned   |
| ---------- | --------- | ------- | ------- |
| local      | 0.62s     | 1.96s   | 0.64s   |
| --no-local | 0.64s     | 2.11s   | 0.89s   |


Part of https://github.com/chapel-lang/chapel/issues/11852
Related to https://github.com/chapel-lang/chapel/issues/11853