@ronawho 2019-06-03
This needs to pass as a correctness test (`start_test test/memory/lifetime/allocation.chpl`) and should run in well under a second for that mode. As a general rule of thumb I try to keep microbenchmarks under 20s execution time too, and this takes 100+ seconds on our main perf box.