@ronawho 2019-06-03
> Can I have an example of a .good file that is a python script and just accepts if it contains three lines? (I.E a regular expression)? I can't easily find any and there isn't any documentation at the ready (that I can find) that tell me how to write such a script.

You can have executable .good files, but that seems like overkill. Typically we just have a .perfexecopts that throw `--perf` or `--printTimings` or something. See https://github.com/chapel-lang/chapel/pull/13009, https://github.com/chapel-lang/chapel/pull/12440, or https://github.com/chapel-lang/chapel/pull/13145 for recent examples.