@ronawho 2019-06-03
> A binary tree benchmark seems a bit interesting. I'll see if I can do that next.

I'd save that for a different PR. And to be clear that issue is talking about the shootout binary trees (in test/release/examples/shootout or test/studies/shootout.) It's still a microbenchmark but it has slightly more interesting allocation patterns.