@ronawho 2019-06-04
FYI here's the timings I see on our main perf box:

| Unmanaged | Shared  | Owned   |
| --------- | ------- | ------- |
| 0.62s     | 1.96s   | 0.64s   |