@ronawho 2019-06-04
Timings on my mac are similar: 

| Unmanaged | Shared  | Owned   |
| --------- | ------- | ------- |
| 0.84s     | 2.15s   | 0.85s   |