@ronawho 2019-06-04
I would expect owned and shared to be relatively close, but it could depend on the backend. Can you update the PR message to whatever you want to be merged? And add something like the following to the end:

```
Timings on Louis's laptop:
// whatever said timings are (maybe run with --numtrials 5 to sanity check)

Timings on chapcs07:

| Unmanaged | Shared  | Owned   |
| --------- | ------- | ------- |
| 0.62s     | 1.96s   | 0.64s   |

Part of https://github.com/chapel-lang/chapel/issues/11852
Related to https://github.com/chapel-lang/chapel/issues/11853
```