@ronawho 2019-06-04
We have a configuration that tests `--no-local` already (https://chapel-lang.org/perf/chapcs/no-local/)