@ronawho 2019-06-04
I see slower timings for --no-local as well:

| config     | Unmanaged | Shared  | Owned   |
| ---------- | --------- | ------- | ------- |
| local      | 0.62s     | 1.96s   | 0.64s   |
| --no-local | 0.64s     | 2.11s   | 0.89s   |