@ronawho 2019-06-04
👍 good perf test -- definitely good to know about the `--no-local` overhead of `owned` and the general overhead of `shared`